requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
--find-links https://download.pytorch.org/whl/torch_stable.html
faiss-cpu>=1.7.4
python-multipart>=0.0.6
nltk>=3.8.1
textblob>=0.17.1
langdetect>=1.0.9 


# unibotenv/bin/activate && pip install -r requirements.txt
#cd src/api && uvicorn main:app --host 0.0.0.0 --port 8001

from transformers import AutoTokenizer, AutoModel
from sentence_transformers import SentenceTransformer

# Load the model using safetensors
model = AutoModel.from_pretrained(
    'models/sentence-transformers/sentence-transformers_all-MiniLM-L6-v2',
    use_safetensors=True
)